Add StartedContainer
