+++ core/src/main/java/org/testcontainers/containers/ContainerLifecycleHooks.java
@@ -1,0 +1,3 @@
+package org.testcontainers.containers;
+
+interface ContainerLifecycleHooks {}
