+++ core/src/main/java/org/testcontainers/containers/StartedContainer.java
@@ -1,0 +1,9 @@
+package org.testcontainers.containers;
+
+import org.testcontainers.UnstableAPI;
+
+@UnstableAPI
+interface StartedContainer extends ContainerState, AutoCloseable {
+    @Override
+    void close();
+}
