+++ modules/kafka/src/main/java/org/testcontainers/containers/StartedKafkaContainer.java
@@ -1,0 +1,8 @@
+package org.testcontainers.containers;
+
+import org.testcontainers.UnstableAPI;
+
+@UnstableAPI
+interface StartedKafkaContainer extends StartedContainer {
+    String getBootstrapServers();
+}
