+++ modules/mongodb/src/main/java/org/testcontainers/containers/StartedMongoDBContainer.java
@@ -1,0 +1,5 @@
+package org.testcontainers.containers;
+
+public interface StartedMongoDBContainer extends StartedContainer {
+    String getConnectionString();
+}
